requirements.txt
streamlit
matplotlib
seaborn
urlextract
wordcloud
pandas
emoji
python-dotenv
requests
